+++ Person.Model.ValueObjects/Json/LandLineConverter.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
@@ -9,7 +9,12 @@ namespace Person.Model.ValueObjects.Json
         public override LandLine? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options) =>
             PhoneNumberFactory.CreateLandLine(ref reader);
 
-        public override void Write(Utf8JsonWriter writer, LandLine? value, JsonSerializerOptions options) =>
-            PhoneNumberFactory.Write(writer, value);
+        public override void Write(Utf8JsonWriter writer, LandLine? value, JsonSerializerOptions options)
+        {
+            if (value is null)
+                writer.WriteNullValue();
+            else
+                PhoneNumberFactory.WriteString(writer, value.Value.Raw);
+        }
     }
 }